+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Add Rail Road diagram to README file
 - Support Delete Step
+- Option `allowDeleteWithoutWhereConditions`
 ### Changed
 - Rename FromStep to SelectFromStep
 - Function selectAsteriskFrom() now can take TableAliases not just Tables